made coverage shared
