+++ shared/Cargo.toml
@@ -3,6 +3,10 @@ name = "shared"
 version = "0.1.0"
 edition = "2024"
 
+[features]
+default = []
+codecov = ["serde", "chrono"]
+
 [dependencies]
 anyhow = "1.0.98"
 async-lsp = "0.2.2"
@@ -15,3 +19,5 @@ thiserror = "2.0.12"
 tokio = "1.47.0"
 tree-sitter = "0.25.8"
 walkdir = "2.5.0"
+serde = { version = "1.0.215", features = ["derive"], optional = true }
+chrono = { version = "0.4.38", features = ["serde"], optional = true }
