+++ shared/src/lib.rs
@@ -1,3 +1,5 @@
 pub mod error;
+#[cfg(feature = "codecov")]
+pub mod codecov;
 
 pub use error::{Context, Error, Result};
