+++ standalone/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-shared = { path = "../shared" }
+shared = { path = "../shared", features = ["codecov"] }
 ast = { path = "../ast" , features = []}
 lsp = { path = "../lsp" }
 
@@ -29,3 +29,4 @@ hex = "0.4"
 [features]
 neo4j = ["ast/neo4j"]
 fulltest = ["ast/fulltest"]
+codecov = []
